+++ src/main/java/de/alpharogroup/swing/table/editor/TableCellButtonEditor.java
@@ -29,7 +29,6 @@ import java.awt.Component;
 import javax.swing.DefaultCellEditor;
 import javax.swing.JButton;
 import javax.swing.JCheckBox;
-import javax.swing.JOptionPane;
 import javax.swing.JTable;
 
 import lombok.Getter;
@@ -83,18 +82,9 @@ public class TableCellButtonEditor extends DefaultCellEditor
 	@Override
 	public Object getCellEditorValue()
 	{
-		if (isClicked())
-		{
-			JOptionPane.showMessageDialog(button, "You clicked the button with the value "
-				+ this.value + " in row index " + row + " and in colunm index " + column + ".");
-		}
 		setClicked(false);
-		String text = "";
-		if (getValue() != null)
-		{
-			text = getValue().toString();
-		}
-		return text;
+		onGetCellEditorValue();
+		return onSetText();
 	}
 
 	/**
@@ -117,15 +107,11 @@ public class TableCellButtonEditor extends DefaultCellEditor
 			getButton().setForeground(table.getForeground());
 			getButton().setBackground(table.getBackground());
 		}
-		String text = "";
-		if (value != null)
-		{
-			text = getValue().toString();
-		}
-		getButton().setText(text);
+		getButton().setText(onSetText());
 		setClicked(true);
 		return getButton();
 	}
+
 
 	/**
 	 * Callback method to interact when the button is clicked.
@@ -133,6 +119,26 @@ public class TableCellButtonEditor extends DefaultCellEditor
 	protected void onClick()
 	{
 		fireEditingStopped();
+	}
+
+	/**
+	 * Callback method to interact in the method getCellEditorValue
+	 */
+	protected void onGetCellEditorValue()
+	{
+	}
+
+	/**
+	 * Callback method to interact when the text is set
+	 */
+	protected String onSetText()
+	{
+		String text = "";
+		if (getValue() != null)
+		{
+			text = getValue().toString();
+		}
+		return text;
 	}
 
 	/**