+++ CHANGELOG.md
@@ -7,6 +7,7 @@ Version 5.1-SNAPSHOT
 ADDED:
  
 - new method created that create a configuration directory file object with factory method that can be overwritten
+- update of email-tails version to 5.0.2
 - update of file-worker version to 5.0.2
 - update of jcommons-lang version to 5.1.1